Добавляет презентер фильтра, переименоввывет файлы фильтра
