+++ src/view/new-filter-view.js
@@ -1,41 +1,27 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import {SortType} from '../utils.js';
 
-const createNewFilterTemplate = () => (
-  `<ul class="sort">
-    <li><a href="#" class="sort__button" data-sort-type="${SortType.DEFAULT}">Sort by default</a></li>
-    <li><a href="#" class="sort__button" data-sort-type="${SortType.SORT_DATE}">Sort by date</a></li>
-    <li><a href="#" class="sort__button" data-sort-type="${SortType.SORT_RATING}">Sort by rating</a></li>
-  </ul>`
-);
+const createNewFilterTemplate = (movies) => {
+  const inWatchlist = movies.filter((movie) => movie['userDetails']['watchlist']);
+  const inHistory = movies.filter((movie) => movie['userDetails']['alreadyWatched']);
+  const inFavorites = movies.filter((movie) => movie['userDetails']['favorite']);
+
+  return (`<nav class="main-navigation">
+    <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
+    <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">${inWatchlist.length}</span></a>
+    <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">${inHistory.length}</span></a>
+    <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">${inFavorites.length}</span></a>
+  </nav>`);
+};
 
 export default class NewFilterView extends AbstractView {
 
-  get template() {
-    return createNewFilterTemplate();
+  constructor(movies) {
+    super();
+    this.movies = movies;
   }
 
-  setClickSortHandler = (callback) => {
-    this._callback.clickSortButton = callback;
-    this.element.addEventListener('click', this.#clickSortHandler);
-  };
-
-  #clickSortHandler = (evt) => {
-    if (evt.target.tagName !== 'A') {
-      return;
-    }
-    evt.preventDefault();
-
-    if (evt.target.classList.contains('sort__button--active')) {
-      evt.target.classList.remove('sort__button--active');
-      this._callback.clickSortButton('default');
-    } else {
-      if (this.element.querySelector('.sort__button--active')) {
-        this.element.querySelector('.sort__button--active').classList.remove('sort__button--active');
-      }
-      evt.target.classList.add('sort__button--active');
-      this._callback.clickSortButton(evt.target.dataset.sortType);
-    }
-  };
+  get template() {
+    return createNewFilterTemplate(this.movies);
+  }
 
 }
